+++ rpc/backend/tracing.go
@@ -26,6 +26,7 @@ import (
 	rpctypes "github.com/evmos/ethermint/rpc/types"
 	ethermint "github.com/evmos/ethermint/types"
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
+	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
 	"github.com/pkg/errors"
 )
 
@@ -117,6 +118,12 @@ func (b *Backend) TraceTransaction(hash common.Hash, config *rpctypes.TraceConfi
 		// 0 is a special value in `ContextWithHeight`
 		contextHeight = 1
 	}
+	// Get basefee from transaction height
+	res, err := b.queryClient.FeeMarket.Params(rpctypes.ContextWithHeight(transaction.Height), &feemarkettypes.QueryParamsRequest{})
+	if err != nil {
+		return nil, err
+	}
+	traceTxRequest.BaseFee = &res.Params.BaseFee
 	traceResult, err := b.queryClient.TraceTx(rpctypes.ContextWithHeight(contextHeight), &traceTxRequest)
 	if err != nil {
 		return nil, err
@@ -231,7 +238,7 @@ func (b *Backend) TraceCall(
 	if err != nil {
 		return nil, err
 	}
-	blk, err := b.TendermintBlockByNumber(blockNr)
+	header, err := b.TendermintHeaderByNumber(blockNr)
 	if err != nil {
 		// the error message imitates geth behavior
 		return nil, errors.New("header not found")
@@ -240,10 +247,10 @@ func (b *Backend) TraceCall(
 	traceCallRequest := evmtypes.QueryTraceCallRequest{
 		Args:            bz,
 		GasCap:          b.RPCGasCap(),
-		ProposerAddress: sdk.ConsAddress(blk.Block.ProposerAddress),
-		BlockNumber:     blk.Block.Height,
-		BlockHash:       common.Bytes2Hex(blk.BlockID.Hash),
-		BlockTime:       blk.Block.Time,
+		ProposerAddress: sdk.ConsAddress(header.Header.ProposerAddress),
+		BlockNumber:     header.Header.Height,
+		BlockHash:       common.Bytes2Hex(header.Header.Hash()),
+		BlockTime:       header.Header.Time,
 		ChainId:         b.chainID.Int64(),
 	}
 
@@ -252,7 +259,7 @@ func (b *Backend) TraceCall(
 	}
 
 	// get the context of provided block
-	contextHeight := blk.Block.Height
+	contextHeight := header.Header.Height
 	if contextHeight < 1 {
 		// 0 is a special value in `ContextWithHeight`
 		contextHeight = 1