+++ x/evm/types/msg.go
@@ -178,11 +178,9 @@ func (msg MsgEthereumTx) ValidateBasic() error {
 	if msg.Data != nil {
 		return errorsmod.Wrapf(errortypes.ErrInvalidRequest, "tx data is deprecated in favor of Raw")
 	}
-
 	if err := msg.Raw.Validate(); err != nil {
 		return err
 	}
-
 	return nil
 }
 
@@ -402,9 +400,4 @@ func (m *MsgUpdateParams) ValidateBasic() error {
 	}
 
 	return m.Params.Validate()
-}
-
-// GetSignBytes implements the LegacyMsg interface.
-func (m MsgUpdateParams) GetSignBytes() []byte {
-	return sdk.MustSortJSON(AminoCdc.MustMarshalJSON(&m))
 }